# internal/ports/openapi_types.go
// Package ports provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.4.1 DO NOT EDIT.
package ports

// ErrorDTO defines model for ErrorDTO.
type ErrorDTO struct {
	Code    string `json:"code"`
	Message string `json:"message"`
}

// ExchangeRateDTO defines model for ExchangeRateDTO.
type ExchangeRateDTO struct {
	From string `json:"from"`
	Rate string `json:"rate"`
	To   string `json:"to"`
}

// GetCryptoExchangeRateParams defines parameters for GetCryptoExchangeRate.
type GetCryptoExchangeRateParams struct {
	From   string  `form:"from" json:"from"`
	To     string  `form:"to" json:"to"`
	Amount float64 `form:"amount" json:"amount"`
}

// GetGlobalExchangeRatesParams defines parameters for GetGlobalExchangeRates.
type GetGlobalExchangeRatesParams struct {
	Currencies string `form:"currencies" json:"currencies"`
}